requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
httpcore==0.17.3
httpx==0.24.1
gunicorn==21.2.0 